+++ internal/commands/extension/uninstall_test.go
@@ -1,0 +1,387 @@
+package extension
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"runtime"
+	"strings"
+	"testing"
+
+	"github.com/indaco/sley/internal/config"
+	"github.com/indaco/sley/internal/testutils"
+	"github.com/urfave/cli/v3"
+)
+
+/* ------------------------------------------------------------------------- */
+/* EXTENSION UNINSTALL COMMAND                                               */
+/* ------------------------------------------------------------------------- */
+
+func TestExtensionUninstallCmd_DeleteFolderVariants(t *testing.T) {
+	extensionName := "mock-extension"
+
+	tests := []struct {
+		name                   string
+		deleteFolder           bool
+		expectDeleted          bool
+		expectParentDirRemoved bool
+	}{
+		{
+			name:                   "delete-folder=false",
+			deleteFolder:           false,
+			expectDeleted:          false,
+			expectParentDirRemoved: false,
+		},
+		{
+			name:                   "delete-folder=true",
+			deleteFolder:           true,
+			expectDeleted:          true,
+			expectParentDirRemoved: true, // only extension, so parent should be removed
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			tmpDir := t.TempDir()
+
+			extensionsRoot := filepath.Join(tmpDir, ".sley-extensions")
+			extensionDir := filepath.Join(extensionsRoot, extensionName)
+			configPath := filepath.Join(tmpDir, ".sley.yaml")
+
+			createExtensionDir(t, extensionDir)
+			writeConfigFile(t, configPath)
+
+			args := []string{"sley", "extension", "uninstall", "--name", extensionName}
+			if tt.deleteFolder {
+				args = append(args, "--delete-folder")
+			}
+
+			cfg := &config.Config{Path: configPath}
+			appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+			output, err := testutils.CaptureStdout(func() {
+				testutils.RunCLITest(t, appCli, args, tmpDir)
+			})
+			if err != nil {
+				t.Fatalf("CLI run failed: %v", err)
+			}
+
+			checkCLIOutput(t, output, extensionName, tt.deleteFolder)
+			checkExtensionDirDeleted(t, extensionDir, tt.expectDeleted)
+			checkExtensionRemovedFromConfig(t, configPath, extensionName)
+
+			// Verify the parent .sley-extensions directory is cleaned up
+			// when it becomes empty after deleting the last extension.
+			checkExtensionDirDeleted(t, extensionsRoot, tt.expectParentDirRemoved)
+		})
+	}
+}
+
+func TestExtensionUninstallCmd_DeleteFolderFailure(t *testing.T) {
+	if runtime.GOOS == "windows" {
+		t.Skip("permission-based RemoveAll failure is unreliable on Windows")
+	}
+
+	tmpDir := t.TempDir()
+	extensionName := "mock-extension"
+	extensionDir := filepath.Join(tmpDir, ".sley-extensions", extensionName)
+	extensionConfigPath := filepath.Join(tmpDir, ".sley.yaml")
+
+	if err := os.MkdirAll(extensionDir, 0755); err != nil {
+		t.Fatalf("failed to create extension directory: %v", err)
+	}
+
+	protectedFile := filepath.Join(extensionDir, "protected.txt")
+	if err := os.WriteFile(protectedFile, []byte("protected"), 0400); err != nil {
+		t.Fatalf("failed to create protected file: %v", err)
+	}
+	if err := os.Chmod(extensionDir, 0500); err != nil {
+		t.Fatalf("failed to chmod extension dir: %v", err)
+	}
+
+	defer func() {
+		_ = os.Chmod(protectedFile, 0600)
+		_ = os.Chmod(extensionDir, 0700)
+	}()
+
+	content := `extensions:
+  - name: mock-extension
+    path: /some/path
+    enabled: true`
+	if err := os.WriteFile(extensionConfigPath, []byte(content), 0644); err != nil {
+		t.Fatalf("failed to write config: %v", err)
+	}
+
+	// Prepare and run the CLI command
+	cfg := &config.Config{Path: extensionConfigPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	var cliErr error
+	_, captureErr := testutils.CaptureStdout(func() {
+		cliErr = testutils.RunCLITestAllowError(t, appCli, []string{
+			"sley", "extension", "uninstall",
+			"--name", extensionName,
+			"--delete-folder",
+		}, tmpDir)
+	})
+	if captureErr != nil {
+		t.Fatalf("failed to capture output: %v", captureErr)
+	}
+
+	_ = os.Chmod(protectedFile, 0600)
+	_ = os.Chmod(extensionDir, 0700)
+
+	if cliErr == nil {
+		t.Fatal("expected error but got nil")
+	}
+
+	expectedMsg := "failed to remove extension directory"
+	if !strings.Contains(cliErr.Error(), expectedMsg) {
+		t.Errorf("expected error message to contain %q, got: %v", expectedMsg, cliErr)
+	}
+
+}
+
+func TestExtensionUninstallCmd_DeleteFolder_ParentDirKeptWhenNotEmpty(t *testing.T) {
+	extensionName := "mock-extension"
+
+	tmpDir := t.TempDir()
+	extensionsRoot := filepath.Join(tmpDir, ".sley-extensions")
+	extensionDir := filepath.Join(extensionsRoot, extensionName)
+	otherExtDir := filepath.Join(extensionsRoot, "other-extension")
+	configPath := filepath.Join(tmpDir, ".sley.yaml")
+
+	// Create both extension directories so parent is not empty after removal.
+	createExtensionDir(t, extensionDir)
+	createExtensionDir(t, otherExtDir)
+	writeConfigFile(t, configPath)
+
+	args := []string{"sley", "extension", "uninstall", "--name", extensionName, "--delete-folder"}
+
+	cfg := &config.Config{Path: configPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	output, err := testutils.CaptureStdout(func() {
+		testutils.RunCLITest(t, appCli, args, tmpDir)
+	})
+	if err != nil {
+		t.Fatalf("CLI run failed: %v", err)
+	}
+
+	checkCLIOutput(t, output, extensionName, true)
+	checkExtensionDirDeleted(t, extensionDir, true)
+
+	// The parent .sley-extensions directory should still exist because
+	// "other-extension" is still present.
+	checkExtensionDirDeleted(t, extensionsRoot, false)
+}
+
+func TestRemoveLocalExtensionsDirIfEmpty(t *testing.T) {
+	tests := []struct {
+		name        string
+		setup       func(t *testing.T) string
+		expectExist bool
+		expectErr   bool
+	}{
+		{
+			name: "removes empty directory",
+			setup: func(t *testing.T) string {
+				t.Helper()
+				dir := filepath.Join(t.TempDir(), ".sley-extensions")
+				if err := os.Mkdir(dir, 0755); err != nil {
+					t.Fatal(err)
+				}
+				return dir
+			},
+			expectExist: false,
+			expectErr:   false,
+		},
+		{
+			name: "keeps non-empty directory",
+			setup: func(t *testing.T) string {
+				t.Helper()
+				dir := filepath.Join(t.TempDir(), ".sley-extensions")
+				child := filepath.Join(dir, "some-extension")
+				if err := os.MkdirAll(child, 0755); err != nil {
+					t.Fatal(err)
+				}
+				return dir
+			},
+			expectExist: true,
+			expectErr:   false,
+		},
+		{
+			name: "no error for non-existent directory",
+			setup: func(t *testing.T) string {
+				t.Helper()
+				return filepath.Join(t.TempDir(), ".sley-extensions")
+			},
+			expectExist: false,
+			expectErr:   false,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			dir := tt.setup(t)
+			err := removeLocalExtensionsDirIfEmpty(dir)
+
+			if tt.expectErr && err == nil {
+				t.Fatal("expected error but got nil")
+			}
+			if !tt.expectErr && err != nil {
+				t.Fatalf("unexpected error: %v", err)
+			}
+
+			_, statErr := os.Stat(dir)
+			exists := statErr == nil
+			if exists != tt.expectExist {
+				t.Errorf("directory exists = %v, want %v", exists, tt.expectExist)
+			}
+		})
+	}
+}
+
+func TestCLI_ExtensionUninstall_MissingName(t *testing.T) {
+	if os.Getenv("TEST_EXTENSION_UNINSTALL_MISSING_NAME") == "1" {
+		tmp := t.TempDir()
+
+		// Write valid .sley.yaml with 1 extension (won't be used, but still required)
+		configPath := filepath.Join(tmp, ".sley.yaml")
+		content := `extensions:
+  - name: mock-extension
+    path: /some/path
+    enabled: true`
+		if err := os.WriteFile(configPath, []byte(content), 0644); err != nil {
+			fmt.Fprintln(os.Stderr, "failed to write config:", err)
+			os.Exit(1)
+		}
+
+		// Prepare and run the CLI command
+		cfg := &config.Config{Path: configPath}
+		appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+		// Run command WITHOUT --name (should trigger the validation)
+		err := appCli.Run(context.Background(), []string{
+			"sley", "extension", "uninstall", "--path", configPath,
+		})
+		if err != nil {
+			fmt.Fprintln(os.Stderr, err)
+			os.Exit(1)
+		}
+
+		// Shouldn't reach here
+		os.Exit(0)
+	}
+
+	cmd := exec.Command(os.Args[0], "-test.run=TestCLI_ExtensionUninstall_MissingName")
+	cmd.Env = append(os.Environ(), "TEST_EXTENSION_UNINSTALL_MISSING_NAME=1")
+	output, err := cmd.CombinedOutput()
+
+	if err == nil {
+		t.Fatal("expected non-zero exit status")
+	}
+
+	expected := "please provide an extension name to uninstall"
+	if !strings.Contains(string(output), expected) {
+		t.Errorf("expected output to contain %q, got:\n%s", expected, output)
+	}
+}
+
+func TestExtensionUninstallCmd_ReadConfigError(t *testing.T) {
+	// The uninstall command reads the config file directly via
+	// RemoveExtension. When the file does not exist the error is
+	// surfaced to the user.
+	tmpDir := t.TempDir()
+	// Do NOT create .sley.yaml so the read fails.
+
+	cfg := &config.Config{Path: filepath.Join(tmpDir, ".sley.yaml")}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	// Capture the output
+	output, err := testutils.CaptureStdout(func() {
+		testutils.RunCLITest(t, appCli, []string{"sley", "extension", "uninstall", "--name", "mock-plugin"}, tmpDir)
+	})
+
+	if err != nil {
+		t.Fatalf("Failed to capture stdout: %v", err)
+	}
+
+	// Expected error message from the file read failure
+	expected := "failed to uninstall extension"
+	if !strings.Contains(output, expected) {
+		t.Errorf("expected output to contain %q, got %q", expected, output)
+	}
+}
+
+func TestExtensionUninstallCmd_ExtensionNotFound(t *testing.T) {
+	tmpDir := t.TempDir()
+	extensionConfigPath := filepath.Join(tmpDir, ".sley.yaml")
+
+	// Create a dummy .sley.yaml configuration file with no extension
+	content := `extensions: []`
+	if err := os.WriteFile(extensionConfigPath, []byte(content), 0644); err != nil {
+		t.Fatalf("failed to create .sley.yaml: %v", err)
+	}
+
+	// Prepare and run the CLI command
+	cfg := &config.Config{Path: extensionConfigPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	// Capture the output
+	output, err := testutils.CaptureStdout(func() {
+		testutils.RunCLITest(t, appCli, []string{"sley", "extension", "uninstall", "--name", "mock-extension"}, tmpDir)
+	})
+
+	if err != nil {
+		t.Fatalf("Failed to capture stdout: %v", err)
+	}
+
+	// Expected error message
+	expected := "extension \"mock-extension\" not found"
+	if !strings.Contains(output, expected) {
+		t.Errorf("expected output to contain %q, got %q", expected, output)
+	}
+}
+
+func TestExtensionUninstallCmd_WriteConfigError(t *testing.T) {
+	if runtime.GOOS == "windows" {
+		t.Skip("permission-based write failure is unreliable on Windows")
+	}
+
+	tmpDir := t.TempDir()
+	extensionConfigPath := filepath.Join(tmpDir, ".sley.yaml")
+
+	// Create a config file then make it read-only so the write fails.
+	content := `extensions:
+  - name: mock-extension
+    path: /path/to/extension
+    enabled: true`
+	if err := os.WriteFile(extensionConfigPath, []byte(content), 0444); err != nil {
+		t.Fatalf("failed to create .sley.yaml: %v", err)
+	}
+	t.Cleanup(func() {
+		_ = os.Chmod(extensionConfigPath, 0644)
+	})
+
+	// Prepare and run the CLI command
+	cfg := &config.Config{Path: extensionConfigPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	// Capture the output
+	output, err := testutils.CaptureStdout(func() {
+		testutils.RunCLITest(t, appCli, []string{"sley", "extension", "uninstall", "--name", "mock-extension"}, tmpDir)
+	})
+
+	if err != nil {
+		t.Fatalf("Failed to capture stdout: %v", err)
+	}
+
+	// Expected error message from the file write failure
+	expected := "failed to uninstall extension"
+	if !strings.Contains(output, expected) {
+		t.Errorf("expected output to contain %q, got %q", expected, output)
+	}
+}
