+++ internal/extensionmgr/setextenabled_test.go
@@ -1,0 +1,264 @@
+package extensionmgr
+
+import (
+	"errors"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+
+	"github.com/goccy/go-yaml"
+	"github.com/indaco/sley/internal/config"
+)
+
+// ---------------------------------------------------------------------------
+// SetExtensionEnabled tests
+// ---------------------------------------------------------------------------
+
+func TestSetExtensionEnabled_Disable(t *testing.T) {
+	tmpDir := t.TempDir()
+	configPath := filepath.Join(tmpDir, ".sley.yaml")
+
+	initial := []byte(`path: .version
+extensions:
+  - name: my-ext
+    path: .sley-extensions/my-ext
+    enabled: true
+`)
+	if err := os.WriteFile(configPath, initial, 0644); err != nil {
+		t.Fatal(err)
+	}
+
+	updater := NewDefaultConfigUpdater(&DefaultYAMLMarshaler{})
+	if err := updater.SetExtensionEnabled(configPath, "my-ext", false); err != nil {
+		t.Fatalf("expected success, got: %v", err)
+	}
+
+	data, err := os.ReadFile(configPath)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	var parsed config.Config
+	if err := yaml.Unmarshal(data, &parsed); err != nil {
+		t.Fatalf("failed to unmarshal updated config: %v", err)
+	}
+
+	if len(parsed.Extensions) != 1 {
+		t.Fatalf("expected 1 extension, got %d", len(parsed.Extensions))
+	}
+	if parsed.Extensions[0].Enabled {
+		t.Errorf("expected extension to be disabled, but it is enabled")
+	}
+
+	// Verify the raw YAML contains enabled: false
+	if !strings.Contains(string(data), "enabled: false") {
+		t.Errorf("expected YAML to contain 'enabled: false', got:\n%s", string(data))
+	}
+}
+
+func TestSetExtensionEnabled_Enable(t *testing.T) {
+	tmpDir := t.TempDir()
+	configPath := filepath.Join(tmpDir, ".sley.yaml")
+
+	initial := []byte(`path: .version
+extensions:
+  - name: my-ext
+    path: .sley-extensions/my-ext
+    enabled: false
+`)
+	if err := os.WriteFile(configPath, initial, 0644); err != nil {
+		t.Fatal(err)
+	}
+
+	updater := NewDefaultConfigUpdater(&DefaultYAMLMarshaler{})
+	if err := updater.SetExtensionEnabled(configPath, "my-ext", true); err != nil {
+		t.Fatalf("expected success, got: %v", err)
+	}
+
+	data, err := os.ReadFile(configPath)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	var parsed config.Config
+	if err := yaml.Unmarshal(data, &parsed); err != nil {
+		t.Fatalf("failed to unmarshal updated config: %v", err)
+	}
+
+	if len(parsed.Extensions) != 1 {
+		t.Fatalf("expected 1 extension, got %d", len(parsed.Extensions))
+	}
+	if !parsed.Extensions[0].Enabled {
+		t.Errorf("expected extension to be enabled, but it is disabled")
+	}
+
+	// Verify the raw YAML contains enabled: true
+	if !strings.Contains(string(data), "enabled: true") {
+		t.Errorf("expected YAML to contain 'enabled: true', got:\n%s", string(data))
+	}
+}
+
+func TestSetExtensionEnabled_NotFound(t *testing.T) {
+	tmpDir := t.TempDir()
+	configPath := filepath.Join(tmpDir, ".sley.yaml")
+
+	initial := []byte("path: .version\nextensions: []\n")
+	if err := os.WriteFile(configPath, initial, 0644); err != nil {
+		t.Fatal(err)
+	}
+
+	updater := NewDefaultConfigUpdater(&DefaultYAMLMarshaler{})
+	err := updater.SetExtensionEnabled(configPath, "nonexistent", true)
+	if err == nil {
+		t.Fatal("expected error for nonexistent extension, got nil")
+	}
+	if !strings.Contains(err.Error(), "not found in configuration") {
+		t.Errorf("expected error to contain 'not found in configuration', got: %v", err)
+	}
+}
+
+func TestSetExtensionEnabled_PreservesComments(t *testing.T) {
+	tmpDir := t.TempDir()
+	configPath := filepath.Join(tmpDir, ".sley.yaml")
+
+	initial := []byte(`# sley configuration file
+# Generated by sley init
+path: .version # path to the version file
+
+# Plugin settings
+plugins:
+  commit-parser: true
+
+# Installed extensions
+extensions:
+  - name: ext-one
+    path: .sley-extensions/ext-one
+    enabled: true
+  - name: ext-two
+    path: .sley-extensions/ext-two
+    enabled: true
+
+# Pre-release hooks
+pre-release-hooks:
+  - changelog:
+      command: git-chglog -o CHANGELOG.md
+`)
+	if err := os.WriteFile(configPath, initial, 0644); err != nil {
+		t.Fatal(err)
+	}
+
+	updater := NewDefaultConfigUpdater(&DefaultYAMLMarshaler{})
+	if err := updater.SetExtensionEnabled(configPath, "ext-one", false); err != nil {
+		t.Fatalf("expected success, got: %v", err)
+	}
+
+	data, err := os.ReadFile(configPath)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	content := string(data)
+
+	// Verify all comments are preserved
+	expectedComments := []string{
+		"# sley configuration file",
+		"# Generated by sley init",
+		"# path to the version file",
+		"# Plugin settings",
+		"# Installed extensions",
+		"# Pre-release hooks",
+	}
+	for _, comment := range expectedComments {
+		if !strings.Contains(content, comment) {
+			t.Errorf("expected comment %q to be preserved, but it is missing.\nActual content:\n%s", comment, content)
+		}
+	}
+
+	// Verify the extension was toggled correctly by parsing the YAML
+	var parsed config.Config
+	if err := yaml.Unmarshal(data, &parsed); err != nil {
+		t.Fatalf("failed to unmarshal updated config: %v", err)
+	}
+
+	if len(parsed.Extensions) != 2 {
+		t.Fatalf("expected 2 extensions, got %d", len(parsed.Extensions))
+	}
+
+	// ext-one should be disabled
+	if parsed.Extensions[0].Name != "ext-one" || parsed.Extensions[0].Enabled {
+		t.Errorf("expected ext-one to be disabled, got: %+v", parsed.Extensions[0])
+	}
+
+	// ext-two should remain enabled
+	if parsed.Extensions[1].Name != "ext-two" || !parsed.Extensions[1].Enabled {
+		t.Errorf("expected ext-two to remain enabled, got: %+v", parsed.Extensions[1])
+	}
+
+	// Verify other sections remain intact
+	if parsed.Path != ".version" {
+		t.Errorf("expected path to be .version, got %q", parsed.Path)
+	}
+	if len(parsed.PreReleaseHooks) != 1 {
+		t.Errorf("expected 1 pre-release hook, got %d", len(parsed.PreReleaseHooks))
+	}
+}
+
+func TestSetExtensionEnabled_ReadFileError(t *testing.T) {
+	invalidPath := filepath.Join(t.TempDir(), "nonexistent.yaml")
+
+	updater := NewDefaultConfigUpdater(&DefaultYAMLMarshaler{})
+	err := updater.SetExtensionEnabled(invalidPath, "test", true)
+	if err == nil || !errors.Is(err, os.ErrNotExist) {
+		t.Fatalf("expected file not found error, got: %v", err)
+	}
+}
+
+func TestSetExtensionEnabled_MarshalError(t *testing.T) {
+	tmp := t.TempDir()
+	configPath := filepath.Join(tmp, ".sley.yaml")
+
+	initial := []byte(`extensions:
+  - name: test-ext
+    path: .sley-extensions/test-ext
+    enabled: true
+`)
+	if err := os.WriteFile(configPath, initial, 0644); err != nil {
+		t.Fatal(err)
+	}
+
+	mockMarshaler := &MockYAMLMarshaler{
+		MarshalFunc: func(v any) ([]byte, error) {
+			return nil, errors.New("forced marshal failure")
+		},
+	}
+	updater := NewDefaultConfigUpdater(mockMarshaler)
+
+	err := updater.SetExtensionEnabled(configPath, "test-ext", false)
+	if err == nil || !strings.Contains(err.Error(), "forced marshal failure") {
+		t.Fatalf("expected marshal error, got: %v", err)
+	}
+}
+
+func TestSetExtensionEnabled_WriteFileError(t *testing.T) {
+	tmp := t.TempDir()
+	configPath := filepath.Join(tmp, ".sley.yaml")
+
+	initial := []byte(`extensions:
+  - name: test-ext
+    path: .sley-extensions/test-ext
+    enabled: true
+`)
+	if err := os.WriteFile(configPath, initial, 0444); err != nil {
+		t.Fatal(err)
+	}
+	t.Cleanup(func() {
+		_ = os.Chmod(configPath, 0644)
+	})
+
+	updater := NewDefaultConfigUpdater(&DefaultYAMLMarshaler{})
+	err := updater.SetExtensionEnabled(configPath, "test-ext", false)
+	if err == nil || !strings.Contains(err.Error(), "permission denied") {
+		t.Fatalf("expected write error, got: %v", err)
+	}
+}
